Changes in the flow of soc.ca solving problems with missing supplementary and passive modalities
A new function that creates an indicator matrix.

The indicator matrix is used in soc.ca to create the matrix for the individuals instead of sup.burt.
This simplifies the flow of the function and gives a performance boost.

result$nid, result$npassive and result$passive have been added to the output.